% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_overbooking.R
\name{run_overbooking}
\alias{run_overbooking}
\title{Launch the Shiny App Included in This Package}
\usage{
run_overbooking()
}
\value{
None. The function launches the app as a side effect.
}
\description{
This function launches the Shiny application bundled within the MATH4753KORBProject1 package.
It assumes the app resides in `inst/app/` and is intended for interactive use.
}
\examples{
\dontrun{
  run_myapp()
}
}
